requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - PostgreSQL
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1

# Vector Database - Pinecone
pinecone-client==2.2.4

# OpenAI
openai==1.3.5

# PDF Processing
PyMuPDF==1.23.8
pdfminer.six==20221105
pypdf==3.17.1

# Text Processing
tiktoken==0.5.1
nltk==3.8.1
langdetect==1.0.9

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring (Optional)
prometheus-client==0.19.0